created ShiftServiceTest.java

+++ LibrarySystem-Backend/src/main/java/ca/mcgill/ecse321/librarysystem/service/ShiftService.java
@@ -62,18 +62,37 @@ public class ShiftService {
 	 
     @Transactional
     public Shift viewSchedule(int shiftID){
-    	Shift shift = shiftRepository.findShiftByShiftID(shiftID); 
-    	return shift;
+		if (shiftRepository.findShiftByShiftID(shiftID) == null){
+			throw new IllegalArgumentException("Shift Id does not exist");
+		}
+		return shiftRepository.findShiftByShiftID(shiftID);
     }
     
     @Transactional
     public Shift getShift(int id) {
-    	Shift shift = shiftRepository.findShiftByShiftID(id);
-    	return shift;
+		if (shiftRepository.findShiftByShiftID(id) == null){
+			throw new IllegalArgumentException("Shift Id does not exist");
+		}
+		return shiftRepository.findShiftByShiftID(id);
     }
     
     @Transactional
     public Shift createShift(int id, DayOfWeek DayOfWeek, Time start, Time end) {
+		if (id == 0){
+            throw new IllegalArgumentException("Shift Id cannot be 0");
+        }
+        if(shiftRepository.findShiftByShiftID(id) != null){
+            throw new IllegalArgumentException("Shift Id already exists");
+        }
+        if (DayOfWeek == null){
+            throw new IllegalArgumentException("Shift day cannot be empty");
+        }
+        if (start == null){
+            throw new IllegalArgumentException("Shift starting time cannot be empty");
+        }
+        if (end == null){
+            throw new IllegalArgumentException("Shift ending time cannot be empty");
+        }
     	Shift shift = new Shift();
     	shift.setDayOfWeek(DayOfWeek);
     	shift.setEndTime(end);
@@ -87,8 +106,11 @@ public class ShiftService {
     }
     
     @Transactional
-    public List<Shift> getShifts(){
-    	List<Shift> shifts=(List<Shift>) shiftRepository.findAll();
+    public Iterable<Shift> getShifts(){
+    	Iterable<Shift> shifts= shiftRepository.findAll();
+		if (((List<Shift>) shifts).size() == 0){
+			throw new IllegalArgumentException("Shift list cannot be empty");
+		}
     	return shifts;
     	
     }
@@ -96,6 +118,9 @@ public class ShiftService {
     @Transactional
     public Shift deleteShift(int id) {
     	Shift shift =shiftRepository.findShiftByShiftID(id);
+		if (shift == null){
+			throw new IllegalArgumentException("Shift Id does not exist");
+		}
     	shiftRepository.delete(shift);
     	return shift;
     }
@@ -107,6 +132,18 @@ public class ShiftService {
     
     @Transactional
     public Shift updateShift(int id, DayOfWeek DayOfWeek, Time start, Time end) {
+		if (id == 0){
+            throw new IllegalArgumentException("Shift Id cannot be 0");
+        }
+        if (DayOfWeek == null){
+            throw new IllegalArgumentException("Shift day cannot be empty");
+        }
+        if (start == null){
+            throw new IllegalArgumentException("Shift starting time cannot be empty");
+        }
+        if (end == null){
+            throw new IllegalArgumentException("Shift ending time cannot be empty");
+        }
     	Shift shift =shiftRepository.findShiftByShiftID(id);
     	shift.setDayOfWeek(DayOfWeek);
     	shift.setEndTime(end);
